Added limits implementation for integrals
